fix(legacy): Fix bootstrapping error.
